Check to make init_env.sh is run from root dir

+++ ci/init_env.sh
@@ -4,9 +4,17 @@ export DJANGO_SETTINGS_MODULE=dream.settings
 # TODO: [ENV-01] Drop database and recreate, remove all migrations and redo them
 # TODO: [ENV-02] install fixtures in initenv.
 
+if [ ! -f manage.py ]; then
+    echo 'Script should be run from project root dir!'
+    exit 1
+fi
+
 # Creates the precommit hook symlink
 cd .git/hooks/
+rm -rf pre-commit
 ln -s ../../ci/git_precommit.sh pre-commit
 
 # Back to dreamframework/
 cd ../../
+
+exit 0
